requirements/base.txt
# Django
django==3.0.7

# Postgres support
psycopg2-binary==2.8.5
psycopg2==2.8.5

# Additional packages
djangorestframework==3.11.0
drf-yasg

isort==5.4.2

django-environ==0.4.5

svgwrite==1.4
